Makefile: Rebuild demo.o whenever queue.h changes
As a student correctly pointed out at the C bootcamp, without this dependency,
the demo module's understanding of the queue module can fall out of sync. For
instance, changing the return type of the deq() function to void will not force
a recompilation of demo.c, which may therefore still attempt to use said
function's return value. The behavior of this is undefined; for instance, on the
x86-64 System V ABI, this will result in a read from %rax, the contents of which
are unspecified and depend on the code emitted by the C compiler when producing
object code for the queue module.